internal: merge object and contract store interfaces
