--- internal/stores/contracts_test.go
@@ -1,485 +1,0 @@
-package stores
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"reflect"
-	"testing"
-
-	rhpv2 "go.sia.tech/core/rhp/v2"
-	"go.sia.tech/core/types"
-	"go.sia.tech/renterd/api"
-	"go.sia.tech/renterd/hostdb"
-	"gorm.io/gorm/schema"
-	"lukechampine.com/frand"
-)
-
-func generateMultisigUC(m, n uint64, salt string) types.UnlockConditions {
-	uc := types.UnlockConditions{
-		PublicKeys:         make([]types.UnlockKey, n),
-		SignaturesRequired: uint64(m),
-	}
-	for i := range uc.PublicKeys {
-		uc.PublicKeys[i].Algorithm = types.SpecifierEd25519
-		uc.PublicKeys[i].Key = frand.Bytes(32)
-	}
-	return uc
-}
-
-// TestSQLContractStore tests SQLContractStore functionality.
-func TestSQLContractStore(t *testing.T) {
-	cs, _, _, err := newTestSQLStore()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create a host for the contract.
-	hk := types.GeneratePrivateKey().PublicKey()
-	err = cs.addTestHost(hk)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Add an announcement.
-	err = cs.insertTestAnnouncement(hk, hostdb.Announcement{NetAddress: "address"})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create random unlock conditions for the host.
-	uc := generateMultisigUC(1, 2, "salt")
-	uc.PublicKeys[1].Key = hk[:]
-	uc.Timelock = 192837
-
-	// Create a contract and set all fields.
-	fcid := types.FileContractID{1, 1, 1, 1, 1}
-	c := rhpv2.ContractRevision{
-		Revision: types.FileContractRevision{
-			ParentID:         fcid,
-			UnlockConditions: uc,
-			FileContract: types.FileContract{
-				RevisionNumber: 200,
-				Filesize:       4096,
-				FileMerkleRoot: types.Hash256{222},
-				WindowStart:    400,
-				WindowEnd:      500,
-				ValidProofOutputs: []types.SiacoinOutput{
-					{
-						Value:   types.NewCurrency64(121),
-						Address: types.Address{2, 1, 2},
-					},
-				},
-				MissedProofOutputs: []types.SiacoinOutput{
-					{
-						Value:   types.NewCurrency64(323),
-						Address: types.Address{2, 3, 2},
-					},
-				},
-				UnlockHash: types.Hash256{6, 6, 6},
-			},
-		},
-		Signatures: [2]types.TransactionSignature{
-			{
-				ParentID:       types.Hash256(fcid),
-				PublicKeyIndex: 0,
-				Timelock:       100000,
-				CoveredFields:  types.CoveredFields{WholeTransaction: true},
-				Signature:      []byte("signature1"),
-			},
-			{
-				ParentID:       types.Hash256(fcid),
-				PublicKeyIndex: 1,
-				Timelock:       200000,
-				CoveredFields:  types.CoveredFields{WholeTransaction: true},
-				Signature:      []byte("signature2"),
-			},
-		},
-	}
-
-	// Look it up. Should fail.
-	ctx := context.Background()
-	_, err = cs.Contract(ctx, c.ID())
-	if !errors.Is(err, ErrContractNotFound) {
-		t.Fatal(err)
-	}
-	contracts, err := cs.ActiveContracts(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(contracts) != 0 {
-		t.Fatalf("should have 0 contracts but got %v", len(contracts))
-	}
-
-	// Insert it.
-	totalCost := types.NewCurrency64(456)
-	startHeight := uint64(100)
-	if _, err := cs.AddContract(ctx, c, totalCost, startHeight); err != nil {
-		t.Fatal(err)
-	}
-
-	// Look it up again.
-	fetched, err := cs.Contract(ctx, c.ID())
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := api.ContractMetadata{
-		ID:          fcid,
-		HostIP:      "address",
-		HostKey:     hk,
-		StartHeight: 100,
-		RenewedFrom: types.FileContractID{},
-		Spending: api.ContractSpending{
-			Uploads:     types.ZeroCurrency,
-			Downloads:   types.ZeroCurrency,
-			FundAccount: types.ZeroCurrency,
-		},
-		TotalCost: totalCost,
-	}
-	if !reflect.DeepEqual(fetched, expected) {
-		t.Fatal("contract mismatch")
-	}
-	contracts, err = cs.ActiveContracts(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(contracts) != 1 {
-		t.Fatalf("should have 1 contracts but got %v", len(contracts))
-	}
-	if !reflect.DeepEqual(contracts[0], expected) {
-		t.Fatal("contract mismatch")
-	}
-
-	// Add a contract set with our contract and assert we can fetch it using the set name
-	if err := cs.SetContractSet(ctx, "foo", []types.FileContractID{contracts[0].ID}); err != nil {
-		t.Fatal(err)
-	}
-	if contracts, err := cs.Contracts(ctx, "foo"); err != nil {
-		t.Fatal(err)
-	} else if len(contracts) != 1 {
-		t.Fatalf("should have 1 contracts but got %v", len(contracts))
-	}
-	if _, err := cs.Contracts(ctx, "bar"); err != ErrContractSetNotFound {
-		t.Fatal(err)
-	}
-
-	// Delete the contract.
-	if err := cs.RemoveContract(ctx, c.ID()); err != nil {
-		t.Fatal(err)
-	}
-
-	// Look it up. Should fail.
-	_, err = cs.Contract(ctx, c.ID())
-	if !errors.Is(err, ErrContractNotFound) {
-		t.Fatal(err)
-	}
-	contracts, err = cs.ActiveContracts(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(contracts) != 0 {
-		t.Fatalf("should have 0 contracts but got %v", len(contracts))
-	}
-
-	// Make sure the db was cleaned up properly through the CASCADE delete.
-	tableCountCheck := func(table interface{}, tblCount int64) error {
-		var count int64
-		if err := cs.db.Model(table).Count(&count).Error; err != nil {
-			return err
-		}
-		if count != tblCount {
-			return fmt.Errorf("expected %v objects in table %v but got %v", tblCount, table.(schema.Tabler).TableName(), count)
-		}
-		return nil
-	}
-	if err := tableCountCheck(&dbContract{}, 0); err != nil {
-		t.Fatal(err)
-	}
-
-	// Check join table count as well.
-	var count int64
-	if err := cs.db.Table("contract_sectors").Count(&count).Error; err != nil {
-		t.Fatal(err)
-	}
-	if count != 0 {
-		t.Fatalf("expected %v objects in contract_sectors but got %v", 0, count)
-	}
-}
-
-// TestRenewContract is a test for AddRenewedContract.
-func TestRenewedContract(t *testing.T) {
-	cs, _, _, err := newTestSQLStore()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create a host for the contract.
-	hk := types.GeneratePrivateKey().PublicKey()
-	err = cs.addTestHost(hk)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Add an announcement.
-	err = cs.insertTestAnnouncement(hk, hostdb.Announcement{NetAddress: "address"})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create random unlock conditions for the host.
-	uc := generateMultisigUC(1, 2, "salt")
-	uc.PublicKeys[1].Key = hk[:]
-	uc.Timelock = 192837
-
-	// Insert a contract.
-	fcid := types.FileContractID{1, 1, 1, 1, 1}
-	c := rhpv2.ContractRevision{
-		Revision: types.FileContractRevision{
-			ParentID:         fcid,
-			UnlockConditions: uc,
-			FileContract: types.FileContract{
-				Filesize:       1,
-				WindowStart:    2,
-				WindowEnd:      3,
-				RevisionNumber: 4,
-			},
-		},
-	}
-	oldContractTotal := types.NewCurrency64(111)
-	oldContractStartHeight := uint64(100)
-	ctx := context.Background()
-	added, err := cs.AddContract(ctx, c, oldContractTotal, oldContractStartHeight)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Assert the contract is returned.
-	if added.RenewedFrom != (types.FileContractID{}) {
-		t.Fatal("unexpected")
-	}
-
-	// Renew it.
-	fcid2 := types.FileContractID{2, 2, 2, 2, 2}
-	renewed := rhpv2.ContractRevision{
-		Revision: types.FileContractRevision{
-			ParentID:         fcid2,
-			UnlockConditions: uc,
-			FileContract: types.FileContract{
-				MissedProofOutputs: []types.SiacoinOutput{},
-				ValidProofOutputs:  []types.SiacoinOutput{},
-			},
-		},
-	}
-	newContractTotal := types.NewCurrency64(222)
-	newContractStartHeight := uint64(200)
-	if _, err := cs.AddRenewedContract(ctx, renewed, newContractTotal, newContractStartHeight, fcid); err != nil {
-		t.Fatal(err)
-	}
-
-	// Contract should be gone from active contracts.
-	_, err = cs.Contract(ctx, fcid)
-	if !errors.Is(err, ErrContractNotFound) {
-		t.Fatal(err)
-	}
-
-	// New contract should exist.
-	newContract, err := cs.Contract(ctx, fcid2)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := api.ContractMetadata{
-		ID:          fcid2,
-		HostIP:      "address",
-		HostKey:     hk,
-		StartHeight: newContractStartHeight,
-		RenewedFrom: fcid,
-		Spending: api.ContractSpending{
-			Uploads:     types.ZeroCurrency,
-			Downloads:   types.ZeroCurrency,
-			FundAccount: types.ZeroCurrency,
-		},
-		TotalCost: newContractTotal,
-	}
-	if !reflect.DeepEqual(newContract, expected) {
-		t.Fatal("mismatch")
-	}
-
-	// Archived contract should exist.
-	var ac dbArchivedContract
-	err = cs.db.Model(&dbArchivedContract{}).
-		Where("fcid", fileContractID(fcid)).
-		Take(&ac).
-		Error
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ac.Model = Model{}
-	expectedContract := dbArchivedContract{
-		FCID:                fileContractID(fcid),
-		Host:                publicKey(c.HostKey()),
-		RenewedTo:           fileContractID(fcid2),
-		Reason:              archivalReasonRenewed,
-		StartHeight:         100,
-		UploadSpending:      zeroCurrency,
-		DownloadSpending:    zeroCurrency,
-		FundAccountSpending: zeroCurrency,
-	}
-	if !reflect.DeepEqual(ac, expectedContract) {
-		fmt.Println(ac)
-		fmt.Println(expectedContract)
-		t.Fatal("mismatch")
-	}
-
-	// Renew it once more.
-	fcid3 := types.FileContractID{3, 3, 3, 3, 3}
-	renewed = rhpv2.ContractRevision{
-		Revision: types.FileContractRevision{
-			ParentID:         fcid3,
-			UnlockConditions: uc,
-			FileContract: types.FileContract{
-				MissedProofOutputs: []types.SiacoinOutput{},
-				ValidProofOutputs:  []types.SiacoinOutput{},
-			},
-		},
-	}
-	newContractTotal = types.NewCurrency64(333)
-	newContractStartHeight = uint64(300)
-
-	// Assert the renewed contract is returned
-	renewedContract, err := cs.AddRenewedContract(ctx, renewed, newContractTotal, newContractStartHeight, fcid2)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if renewedContract.RenewedFrom != fcid2 {
-		t.Fatal("unexpected")
-	}
-}
-
-// TestAncestorsContracts verifies that AncestorContracts returns the right
-// ancestors in the correct order.
-func TestAncestorsContracts(t *testing.T) {
-	cs, _, _, err := newTestSQLStore()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	hk := types.PublicKey{1, 2, 3}
-	if err := cs.addTestHost(hk); err != nil {
-		t.Fatal(err)
-	}
-
-	// Create a chain of 4 contracts.
-	// Their start heights are 0, 1, 2, 3.
-	fcids := []types.FileContractID{{1}, {2}, {3}, {4}}
-	if _, err := cs.addTestContract(fcids[0], hk); err != nil {
-		t.Fatal(err)
-	}
-	for i := 1; i < len(fcids); i++ {
-		if _, err := cs.addTestRenewedContract(fcids[i], fcids[i-1], hk, uint64(i)); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// Fetch the ancestors but only the ones with a startHeight >= 1. That
-	// should return 2 contracts. The active one with height 3 isn't
-	// returned and the one with height 0 is also not returned.
-	contracts, err := cs.AncestorContracts(context.Background(), fcids[len(fcids)-1], 1)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(contracts) != len(fcids)-2 {
-		t.Fatal("wrong number of contracts returned", len(contracts))
-	}
-	for i := 0; i < len(contracts)-1; i++ {
-		if !reflect.DeepEqual(contracts[i], api.ArchivedContract{
-			ID:        fcids[len(fcids)-2-i],
-			HostKey:   hk,
-			RenewedTo: fcids[len(fcids)-1-i],
-		}) {
-			t.Fatal("wrong contract", i)
-		}
-	}
-}
-
-func (s *SQLStore) addTestContracts(keys []types.PublicKey) (fcids []types.FileContractID, contracts []api.ContractMetadata, err error) {
-	cnt, err := s.contractsCount()
-	if err != nil {
-		return nil, nil, err
-	}
-	for i, key := range keys {
-		fcids = append(fcids, types.FileContractID{byte(int(cnt) + i + 1)})
-		contract, err := s.addTestContract(fcids[len(fcids)-1], key)
-		if err != nil {
-			return nil, nil, err
-		}
-		contracts = append(contracts, contract)
-	}
-	return
-}
-
-func (s *SQLStore) addTestContract(fcid types.FileContractID, hk types.PublicKey) (api.ContractMetadata, error) {
-	rev := testContractRevision(fcid, hk)
-	return s.AddContract(context.Background(), rev, types.ZeroCurrency, 0)
-}
-
-func (s *SQLStore) addTestRenewedContract(fcid, renewedFrom types.FileContractID, hk types.PublicKey, startHeight uint64) (api.ContractMetadata, error) {
-	rev := testContractRevision(fcid, hk)
-	return s.AddRenewedContract(context.Background(), rev, types.ZeroCurrency, startHeight, renewedFrom)
-}
-
-func (s *SQLStore) contractsCount() (cnt int64, err error) {
-	err = s.db.
-		Model(&dbContract{}).
-		Count(&cnt).
-		Error
-	return
-}
-
-func testContractRevision(fcid types.FileContractID, hk types.PublicKey) rhpv2.ContractRevision {
-	uc := generateMultisigUC(1, 2, "salt")
-	uc.PublicKeys[1].Key = hk[:]
-	uc.Timelock = 192837
-	return rhpv2.ContractRevision{
-		Revision: types.FileContractRevision{
-			ParentID:         fcid,
-			UnlockConditions: uc,
-			FileContract: types.FileContract{
-				RevisionNumber: 200,
-				Filesize:       4096,
-				FileMerkleRoot: types.Hash256{222},
-				WindowStart:    400,
-				WindowEnd:      500,
-				ValidProofOutputs: []types.SiacoinOutput{
-					{
-						Value:   types.NewCurrency64(121),
-						Address: types.Address{2, 1, 2},
-					},
-				},
-				MissedProofOutputs: []types.SiacoinOutput{
-					{
-						Value:   types.NewCurrency64(323),
-						Address: types.Address{2, 3, 2},
-					},
-				},
-				UnlockHash: types.Hash256{6, 6, 6},
-			},
-		},
-		Signatures: [2]types.TransactionSignature{
-			{
-				ParentID:       types.Hash256(fcid),
-				PublicKeyIndex: 0,
-				Timelock:       100000,
-				CoveredFields:  types.CoveredFields{WholeTransaction: true},
-				Signature:      []byte("signature1"),
-			},
-			{
-				ParentID:       types.Hash256(fcid),
-				PublicKeyIndex: 1,
-				Timelock:       200000,
-				CoveredFields:  types.CoveredFields{WholeTransaction: true},
-				Signature:      []byte("signature2"),
-			},
-		},
-	}
-}
