+++ internal/stores/sql.go
@@ -71,10 +71,15 @@ func NewSQLStore(conn gorm.Dialector, migrate bool, persistInterval time.Duratio
 	if migrate {
 		// Create the tables.
 		tables := []interface{}{
-			// bus.ContractStore tables
+			// bus.MetadataStore tables
 			&dbArchivedContract{},
 			&dbContract{},
 			&dbContractSet{},
+			&dbObject{},
+			&dbSector{},
+			&dbShard{},
+			&dbSlab{},
+			&dbSlice{},
 
 			// bus.HostDB tables
 			&dbAnnouncement{},
@@ -82,13 +87,6 @@ func NewSQLStore(conn gorm.Dialector, migrate bool, persistInterval time.Duratio
 			&dbHost{},
 			&dbInteraction{},
 			&dbBlocklistEntry{},
-
-			// bus.ObjectStore tables
-			&dbObject{},
-			&dbSlice{},
-			&dbSlab{},
-			&dbSector{},
-			&dbShard{},
 
 			// bus.SettingStore tables
 			&dbSetting{},
